+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
 
+## Unreleased
+### Added
+- More Dialogs to show installation status
+- Animated Toolbar Icons to show progress
+
+### Changed
+- Open Ungit Viewer only after startup has completed
+
+
 ## 0.2.0 - 2015-07-29
 ### Added
 - Self-installer for npm dependencies